renamed opencv_map_maker to map_maker for clarity
